more sanity checking

+++ www/store.php
@@ -43,27 +43,52 @@
 	switch ($_SERVER['REQUEST_METHOD']) {
 		case 'PUT':
 			if (!file_exists($dirname)) {
-				mkdir($dirname, true);
+				// suppress errors from mkdir, it is noisy
+				$res = @mkdir($dirname, true);
+				if ($res == false) {
+					header($protocol . " 412 Precondition Failed");
+					echo "Directory not writeable\n";
+					exit();
+				}
 			}
 
-			if ($filename) {
-				/* PUT data comes in on the stdin stream */
-				$in = fopen("php://input", "r");
+			if ( !is_writable($dirname) ){
+				header($protocol . " 412 Precondition Failed");
+				echo "Directory not writeable\n";
+				exit();
+			} else {
+				if ( $filename ) {
+					$exists = file_exists($dirname.$filename);
+					if (
+						($exists === true && is_writable($dirname.$filename) ) ||
+						$exists === false
+					){
+						/* PUT data comes in on the stdin stream */
+						$in = fopen("php://input", "r");
 
-				/* Open a file for writing */
-				$tempfile = tempnam($dirname, 'put-XXXXXX');
+						/* Open a file for writing */
+						$tempfile = tempnam($dirname, 'put-XXXXXX');
 
-				$out = fopen($tempfile, "w");
-				$res = stream_copy_to_stream($in,$out);
+						$out = fopen($tempfile, "w");
+						$res = stream_copy_to_stream($in,$out);
 
-				/* Close the streams */
-				fclose($out);
-				fclose($in);
+						/* Close the streams */
+						fclose($out);
+						fclose($in);
 
-				if($res) {
-					rename($tempfile, $dirname.$filename);
-				} else {
-					unlink($tempfile);
+						if($res) {
+							$res = rename($tempfile, $dirname.$filename);
+							if ($res == false) {
+								header($protocol . " 412 Precondition Failed");
+								unlink($tempfile);
+							}
+						} else {
+							unlink($tempfile);
+						}
+					} else {
+						header($protocol . " 412 Precondition Failed");
+						echo "Target file not writeable\n";
+					}
 				}
 			}
 			break;
